feat (Table.tsx) : 초대하기 버튼 생성
