+++ components/Table/Table.module.css
@@ -39,6 +39,21 @@
   background-color: var(--gray30);
 }
 
+.invitebutton {
+  width: 10rem;
+  height: 3.2rem;
+  background-color: var(--violet);
+  color: var(--white);
+  font-size: 1.4rem;
+  font-weight: 500;
+  border-radius: 0.4rem;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 0.8rem;
+  grid-area: invite;
+}
+
 .tableindex {
   display: grid;
 }
